requirements.txt
attrs==23.1.0
aws-cdk-lib==2.84.0
aws-cdk.asset-awscli-v1==2.2.196
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v5==2.0.165
blinker==1.6.2
cachetools==5.3.1
cattrs==23.1.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
constructs==10.2.54
cryptography==41.0.1
ecdsa==0.18.0
exceptiongroup==1.1.1
Flask==2.3.2
google-auth==2.20.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsii==1.84.0
jwt==1.3.1
MarkupSafe==2.1.3
publication==0.0.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
typeguard==2.13.3
typing_extensions==4.6.3
urllib3==1.26.16
Werkzeug==2.3.6
zipp==3.15.0
